requirements.txt
click==8.1.3
dash==2.11.1
dash_bio==1.0.2
hdbscan==0.8.30
joblib==1.3.1
numpy==1.24.4
plotly==5.15.0
scikit_learn==1.3.0
tqdm==4.65.0
umap_learn==0.5.3
